MP2/requirements.txt
tqdm
numpy
opencv-python
scipy
scikit-image
matplotlib
absl-py
pillow